fix mistake v3

+++ myteam_spring/src/main/java/by/example/myteam/dao/PersonDAO.java
@@ -10,7 +10,7 @@ public interface PersonDAO {
 
     void savePerson(Person person);
 
-    Person getPerson(long id);
+    Person getPerson(int id);
 
-    void deletePerson(long id);
+    void deletePerson(int id);
 }
