+++ myteam_spring/src/main/java/by/example/myteam/dao/PersonDAOImpl.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.Query;
+import java.time.Period;
 import java.util.List;
 
 @Repository
@@ -34,13 +35,13 @@ public class PersonDAOImpl implements PersonDAO {
     }
 
     @Override
-    public Person getPerson(long id) {
+    public Person getPerson(int id) {
         Session session = sessionFactory.getCurrentSession();
         return session.get(Person.class, id);
     }
 
     @Override
-    public void deletePerson(long id) {
+    public void deletePerson(int id) {
         Session session = sessionFactory.getCurrentSession();
         Query query = session.createQuery("DELETE from Person WHERE id = :personId");
         query.setParameter("personId", id);
